refactor: view 폴더 생성
